+++ arangod/Aql/ExecutionNode/RemoveNode.cpp
@@ -28,6 +28,7 @@
 #include "Aql/ExecutionPlan.h"
 #include "Aql/Executor/ModificationExecutor.h"
 #include "Aql/Executor/ModificationExecutorHelpers.h"
+#include "Aql/ModificationExecutorFlags.h"
 #include "Aql/SingleRowFetcher.h"
 #include "Aql/Variable.h"
 #include "Basics/StaticStrings.h"
