+++ arangod/Aql/InsertModifier.cpp
@@ -24,6 +24,7 @@
 #include "InsertModifier.h"
 
 #include "Aql/AqlValue.h"
+#include "Aql/Collection.h"
 #include "Aql/Executor/ModificationExecutor.h"
 #include "Aql/Executor/ModificationExecutorHelpers.h"
 #include "Basics/Common.h"
